card_scanner/requirements.txt
Flask==2.3.3
pytesseract==0.3.10
Pillow==10.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
opencv-python==4.8.0.76
numpy==1.25.2
prometheus-flask-exporter==0.22.4
Flask-Limiter==3.3.1
gunicorn==21.2.0
structlog==23.1.0
Flask-CORS==4.0.0